identity id in DeleteIdentity should be a BigInteger

+++ src/main/java/com/enjin/enjincoin/sdk/model/service/identities/DeleteIdentity.java
@@ -2,9 +2,11 @@ package com.enjin.enjincoin.sdk.model.service.identities;
 
 import com.enjin.enjincoin.sdk.graphql.GraphQLRequest;
 
+import java.math.BigInteger;
+
 public class DeleteIdentity extends GraphQLRequest.Builder {
 
-    public DeleteIdentity withIdentityId(int id) {
+    public DeleteIdentity withIdentityId(BigInteger id) {
         withParameter("id", id);
         return this;
     }
@@ -14,7 +16,7 @@ public class DeleteIdentity extends GraphQLRequest.Builder {
         return this;
     }
 
-    public static DeleteIdentity unlink(int id) {
+    public static DeleteIdentity unlink(BigInteger id) {
         return new DeleteIdentity().withIdentityId(id)
                                    .withUnlink(true);
     }
